+++ circus-train-s3-s3-copier/src/main/java/com/hotels/bdp/circustrain/s3s3copier/aws/JceksAmazonS3ClientFactory.java
@@ -40,99 +40,54 @@ public class JceksAmazonS3ClientFactory implements AmazonS3ClientFactory {
   private final static Logger LOG = LoggerFactory.getLogger(JceksAmazonS3ClientFactory.class);
 
   private final Security security;
-  private final HiveConf sourceHiveConf;
-  private final HiveConf replicaHiveConf;
+  private final HiveConf conf;
 
   @Autowired
-  public JceksAmazonS3ClientFactory(Security security, HiveConf sourceHiveConf, HiveConf replicaHiveConf) {
+  public JceksAmazonS3ClientFactory(Security security, HiveConf replicaHiveConf) {
     this.security = security;
-    this.sourceHiveConf = sourceHiveConf;
-    this.replicaHiveConf = replicaHiveConf;
+    this.conf = replicaHiveConf;
   }
 
   public JceksAmazonS3ClientFactory(Security security) {
     this.security = security;
-    this.sourceHiveConf = new HiveConf();
-    this.replicaHiveConf = new HiveConf();
+    this.conf = new HiveConf();
   }
 
   @Override
   public AmazonS3 newInstance(AmazonS3URI uri, S3S3CopierOptions s3s3CopierOptions) {
     LOG.debug("trying to get a client for uri '{}'", uri);
-    AmazonS3 globalClient;
-//    AmazonS3 globalClient = newGlobalInstance(s3s3CopierOptions);
-    
-    
-    String assumedRole = s3s3CopierOptions.getSourceAssumedRole();
-    LOG.info("HELLO, WORLD! Creating instance with role " + assumedRole);
-    if (assumedRole != null) {
-      Configuration conf = sourceHiveConf;
-      conf.addResource(AssumeRoleCredentialProvider.ASSUME_ROLE_PROPERTY_NAME);
-      conf.set(AssumeRoleCredentialProvider.ASSUME_ROLE_PROPERTY_NAME, assumedRole);
-
-      LOG.info("GLOBAL SOURCE INSTANCE - has conf");
-      globalClient = newGlobalInstance(s3s3CopierOptions, conf);
-    } else {
-      LOG.debug("trying to get a target client for uri '{}'", uri);
-      globalClient = newGlobalInstance(s3s3CopierOptions);
-    }
-    
-    
-    
+    AmazonS3 globalClient = newGlobalInstance(s3s3CopierOptions);
     try {
       String bucketRegion = regionForUri(globalClient, uri);
       LOG.debug("Bucket region: {}", bucketRegion);
-//      return newInstance(bucketRegion, s3s3CopierOptions);
-      return newSourceInstance(bucketRegion, s3s3CopierOptions);
+      return newInstance(bucketRegion, s3s3CopierOptions);
     } catch (IllegalArgumentException e) {
       LOG.warn("Using global (non region specific) client", e);
       return globalClient;
     }
   }
 
-//  public AmazonS3 newSourceInstance(AmazonS3URI uri, S3S3CopierOptions s3s3CopierOptions) {
-//    return newInstance(uri, s3s3CopierOptions);
-//  }
-
   public AmazonS3 newTargetInstance(AmazonS3URI uri, S3S3CopierOptions s3s3CopierOptions) {
     AmazonS3 globalClient;
 
-    // ****** 
-    // Create global instance with conf 
-    String assumedRole = s3s3CopierOptions.getTargetAssumedRole();
-    LOG.info("HELLO, WORLD! Creating instance with role = " + assumedRole);
+    String assumedRole = s3s3CopierOptions.getAssumedRole();
     if (assumedRole != null) {
-      Configuration conf = replicaHiveConf;
-      conf.addResource(AssumeRoleCredentialProvider.ASSUME_ROLE_PROPERTY_NAME);
-      conf.set(AssumeRoleCredentialProvider.ASSUME_ROLE_PROPERTY_NAME, assumedRole);
-
-      LOG.info("GLOBAL TARGET INSTANCE - has conf");
-      globalClient = newGlobalInstance(s3s3CopierOptions, conf);
+      globalClient = newGlobalInstance(s3s3CopierOptions, setConfRole(conf, assumedRole));
     } else {
-      LOG.debug("trying to get a target client for uri '{}'", uri);
       globalClient = newGlobalInstance(s3s3CopierOptions);
     }
-
     try {
-      System.out.println("Going to Sleep for 60 sec...");
-      Thread.sleep(10000);
-
-      String bucketRegion = regionForUri(globalClient, uri); // ********************************
+      String bucketRegion = regionForUri(globalClient, uri);
       LOG.debug("Bucket region: {}", bucketRegion);
       return newTargetInstance(bucketRegion, s3s3CopierOptions);
-    } catch (IllegalArgumentException | InterruptedException e) {
+    } catch (IllegalArgumentException e) {
       LOG.warn("Using global (non region specific) client", e);
       return globalClient;
     }
   }
 
   private String regionForUri(AmazonS3 client, AmazonS3URI uri) {
-    String bucket = uri.getBucket();
-    LOG.info("GOT BUCKET: " + bucket);
-    
-    String bucketRegion = client.getBucketLocation(bucket); // Causing access denied ***
-
-//    String bucketRegion = client.getBucketLocation(uri.getBucket());
+    String bucketRegion = client.getBucketLocation(uri.getBucket());
     Region region = Region.fromValue(bucketRegion);
 
     // S3 doesn't have a US East 1 region, US East 1 is really the region
@@ -150,31 +105,31 @@ public class JceksAmazonS3ClientFactory implements AmazonS3ClientFactory {
     return bucketRegion;
   }
 
+  private AmazonS3 newInstance(String region, S3S3CopierOptions s3s3CopierOptions) {
+    HadoopAWSCredentialProviderChain credentialsChain = getCredentialsProviderChain();
+    AmazonS3ClientBuilder builder = AmazonS3ClientBuilder.standard().withCredentials(credentialsChain);
+    URI s3Endpoint = s3s3CopierOptions.getS3Endpoint(region);
+    if (s3Endpoint != null) {
+      EndpointConfiguration endpointConfiguration = new EndpointConfiguration(s3Endpoint.toString(), region);
+      builder.withEndpointConfiguration(endpointConfiguration);
+    } else {
+      builder.withRegion(region);
+    }
+    return builder.build();
+  }
+
   private AmazonS3 newGlobalInstance(S3S3CopierOptions s3s3CopierOptions) {
     HadoopAWSCredentialProviderChain credentialsChain = getCredentialsProviderChain();
-//    AmazonS3ClientBuilder builder = AmazonS3ClientBuilder.standard().withForceGlobalBucketAccessEnabled(Boolean.TRUE)
-//        .withCredentials(credentialsChain);
-//    URI s3Endpoint = s3s3CopierOptions.getS3Endpoint();
-//    if (s3Endpoint != null) {
-//      EndpointConfiguration endpointConfiguration = new EndpointConfiguration(s3Endpoint.toString(),
-//          Region.US_Standard.getFirstRegionId());
-//      builder.withEndpointConfiguration(endpointConfiguration);
-//    }
-//    return builder.build();
-    
     return buildGlobalInstance(s3s3CopierOptions, credentialsChain);
   }
-  
-  // *****
+
   private AmazonS3 newGlobalInstance(S3S3CopierOptions s3s3CopierOptions, Configuration conf) {
-    // what does the global instance do?
-    // give it conf with the role - target role ????
     HadoopAWSCredentialProviderChain credentialsChain = getCredentialsProviderChain(conf);
     return buildGlobalInstance(s3s3CopierOptions, credentialsChain);
   }
-  
-  // *****
-  private AmazonS3 buildGlobalInstance(S3S3CopierOptions s3s3CopierOptions, HadoopAWSCredentialProviderChain credentialsChain) {
+
+  private AmazonS3 buildGlobalInstance(S3S3CopierOptions s3s3CopierOptions,
+      HadoopAWSCredentialProviderChain credentialsChain) {
     AmazonS3ClientBuilder builder = AmazonS3ClientBuilder.standard().withForceGlobalBucketAccessEnabled(Boolean.TRUE)
         .withCredentials(credentialsChain);
     URI s3Endpoint = s3s3CopierOptions.getS3Endpoint();
@@ -186,50 +141,12 @@ public class JceksAmazonS3ClientFactory implements AmazonS3ClientFactory {
     return builder.build();
   }
 
-  // *****
-  // Removing the role from the source instance
-  // do we want the source client to have the target role, or the target client ?
-
-//  private AmazonS3 newInstance(String region, S3S3CopierOptions s3s3CopierOptions) {
-//    HadoopAWSCredentialProviderChain credentialsChain = getCredentialsProviderChain();
-//    AmazonS3ClientBuilder builder = AmazonS3ClientBuilder.standard().withCredentials(credentialsChain);
-//    URI s3Endpoint = s3s3CopierOptions.getS3Endpoint(region);
-//    if (s3Endpoint != null) {
-//      EndpointConfiguration endpointConfiguration = new EndpointConfiguration(s3Endpoint.toString(), region);
-//      builder.withEndpointConfiguration(endpointConfiguration);
-//    } else {
-//      builder.withRegion(region);
-//    }
-//    return builder.build();
-//  }
-  // *****
-
-  private AmazonS3 newSourceInstance(String region, S3S3CopierOptions s3s3CopierOptions) {
-    HadoopAWSCredentialProviderChain credentialsChain;
-
-    String assumedRole = s3s3CopierOptions.getSourceAssumedRole();
-    LOG.info("HELLO, WORLD! Source role = " + assumedRole);
-    if (assumedRole != null) {
-      Configuration conf = sourceHiveConf;
-      conf.addResource(AssumeRoleCredentialProvider.ASSUME_ROLE_PROPERTY_NAME);
-      conf.set(AssumeRoleCredentialProvider.ASSUME_ROLE_PROPERTY_NAME, assumedRole);
-      credentialsChain = getCredentialsProviderChain(conf);
-    } else {
-      credentialsChain = getCredentialsProviderChain();
-    }
-    return buildClient(region, s3s3CopierOptions, credentialsChain);
-  }
-
   private AmazonS3 newTargetInstance(String region, S3S3CopierOptions s3s3CopierOptions) {
     HadoopAWSCredentialProviderChain credentialsChain;
 
-    String assumedRole = s3s3CopierOptions.getTargetAssumedRole();
-    LOG.info("HELLO, WORLD! Target role = " + assumedRole);
+    String assumedRole = s3s3CopierOptions.getAssumedRole();
     if (assumedRole != null) {
-      Configuration conf = replicaHiveConf;
-      conf.addResource(AssumeRoleCredentialProvider.ASSUME_ROLE_PROPERTY_NAME);
-      conf.set(AssumeRoleCredentialProvider.ASSUME_ROLE_PROPERTY_NAME, assumedRole);
-      credentialsChain = getCredentialsProviderChain(conf);
+      credentialsChain = getCredentialsProviderChain(setConfRole(conf, assumedRole));
     } else {
       credentialsChain = getCredentialsProviderChain();
     }
@@ -260,10 +177,13 @@ public class JceksAmazonS3ClientFactory implements AmazonS3ClientFactory {
   }
 
   private HadoopAWSCredentialProviderChain getCredentialsProviderChain(Configuration conf) {
-    HadoopAWSCredentialProviderChain credentialsChain = null;
-
-    credentialsChain = new HadoopAWSCredentialProviderChain(conf);
-
-    return credentialsChain;
+    return new HadoopAWSCredentialProviderChain(conf);
   }
+
+  private Configuration setConfRole(Configuration conf, String assumedRole) {
+    conf.addResource(AssumeRoleCredentialProvider.ASSUME_ROLE_PROPERTY_NAME);
+    conf.set(AssumeRoleCredentialProvider.ASSUME_ROLE_PROPERTY_NAME, assumedRole);
+    return conf;
+  }
+
 }